# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: VSS Operations API
  version: 0.3.0
  description: Canonical on-premise operations API for VSS/OpenClaw.
servers:
  - url: http://127.0.0.1:3001
components:
  parameters:
    ActorRef:
      name: X-Actor-Ref
      in: header
      required: true
      schema: { type: string, minLength: 1 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8 }
    SourceMessageId:
      name: X-Source-Message-Id
      in: header
      required: false
      schema: { type: string }
  schemas:
...
    EventInput:
      type: object
      required: [name, startOn, endOn, eventOrganizerId, venueId]
      properties:
        name: { type: string }
        alias: { type: [string, 'null'] }
        notes: { type: [string, 'null'] }
        startOn: { type: string, format: date }
        endOn: { type: string, format: date }
        eventOrganizerId: { type: string, format: uuid }
        venueId: { type: string, format: uuid }
...
paths:
...
  /api/v1/event-organizers:
    get: { summary: List active event organizers }
    post:
      summary: Create event organizer
      parameters: [{ $ref: '#/components/parameters/ActorRef' }, { $ref: '#/components/parameters/IdempotencyKey' }, { $ref: '#/components/parameters/SourceMessageId' }]
      responses: { '201': { description: Created } }
...
  /api/v1/events:
    get:
      summary: List events
      parameters:
        - { name: eventOrganizerId, in: query, schema: { type: string, format: uuid } }
        - { name: venueId, in: query, schema: { type: string, format: uuid } }
        - { name: startsFrom, in: query, schema: { type: string, format: date } }
        - { name: startsUntil, in: query, schema: { type: string, format: date } }
        - { name: status, in: query, schema: { enum: [ACTIVE, CANCELLED] } }
      responses: { '200': { description: Paginated event list } }
    post:
      summary: Create event
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/EventInput' } } }
      responses: { '201': { description: Created } }
...